Refactor color config
