+++ app/src/components/ProtvistaFilter.jsx
@@ -5,7 +5,7 @@ import ProtvistaFilter from "protvista-filter";
 import ProtvistaVariationAdapter from "protvista-variation-adapter";
 import ProtvistaManager from "protvista-manager";
 import loadWebComponent from "../utils/load-web-component";
-import filters from "../mocks/filterConfig";
+import filters, {colourConfig} from "../mocks/filterConfig";
 
 const ProtvistaFilterWrapper = () => {
   loadWebComponent("protvista-variation", ProtvistaVariation);
@@ -15,10 +15,14 @@ const ProtvistaFilterWrapper = () => {
   loadWebComponent("protvista-filter", ProtvistaFilter);
 
   const ref = useRef(null);
+  const variationRef = useRef(null);
 
   useEffect(() => {
     if (ref !== null) {
       ref.current.filters = filters;
+    }
+    if(variationRef !== null) {
+      variationRef.current.colourConfig = colourConfig;
     }
   });
 
@@ -33,7 +37,7 @@ const ProtvistaFilterWrapper = () => {
           for="my-variation-track"
           ref={ref}
         />
-        <protvista-variation length="770" id="my-variation-track">
+        <protvista-variation length="770" id="my-variation-track" ref={variationRef}>
           <protvista-variation-adapter>
             <data-loader>
               <source src="https://www.ebi.ac.uk/proteins/api/variation/P05067" />
